pyproject.toml
[project]
name = "frankentui-site"
version = "0.1.0"
description = "A terminal-themed personal website built with FastHTML"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "monsterui>=1.0.32",
    "python-fasthtml>=0.12.35",
    "python-frontmatter>=1.1.0",
]
